Flexmark - refactored code
