+++ src/main/java/org/concordion/internal/parser/flexmark/ConcordionCommandNodeRenderer.java
@@ -13,9 +13,9 @@ import org.concordion.internal.parser.support.Attribute;
 import java.util.HashSet;
 import java.util.Set;
 
-public class ConcordionNodeRenderer implements NodeRenderer {
+public class ConcordionCommandNodeRenderer implements NodeRenderer {
 
-    public ConcordionNodeRenderer(DataHolder options) {
+    public ConcordionCommandNodeRenderer(DataHolder options) {
     }
 
     @Override
@@ -27,28 +27,7 @@ public class ConcordionNodeRenderer implements NodeRenderer {
                 renderCommand(node, context, html, node.getCommand(), node.getExpression());
             }
         }));
-        set.add(new NodeRenderingHandler<ConcordionExampleNode>(ConcordionExampleNode.class, new CustomNodeRenderer<ConcordionExampleNode>() {
-            @Override
-            public void render(ConcordionExampleNode node, NodeRendererContext context, HtmlWriter html) {
-                renderCommand(node, context, html, node.getCommand(), node.getExpression());
-            }
-        }));
-        set.add(new NodeRenderingHandler<ConcordionLinkNode>(ConcordionLinkNode.class, new CustomNodeRenderer<ConcordionLinkNode>() {
-            @Override
-            public void render(ConcordionLinkNode node, NodeRendererContext context, HtmlWriter html) {
-                renderCommand(node, context, html);
-            }
-        }));
-
         return set;
-    }
-
-    private void renderCommand(ConcordionLinkNode node, NodeRendererContext context, HtmlWriter html) {
-        html.attr("href", node.getUrl());
-        html.attr(node.getCommand(), node.getRunner());
-        html.withAttr().tag("a");
-        context.renderChildren(node);
-        html.tag("/a");
     }
 
     private void renderCommand(ConcordionCommandNode node, NodeRendererContext context, HtmlWriter html, String command, String expression) {
@@ -63,23 +42,10 @@ public class ConcordionNodeRenderer implements NodeRenderer {
         html.closeTag("span");
     }
 
-    private void renderCommand(ConcordionExampleNode node, NodeRendererContext context, HtmlWriter html, String command, String expression) {
-        // TODO extract method
-        final Attributes nodeAttributes = new Attributes();
-        nodeAttributes.addValue(command, expression);
-        for (Attribute attribute : node.getAttributes()) {
-            nodeAttributes.addValue(attribute.name, attribute.value);
-        }
-        html.setAttributes(nodeAttributes).withAttr().tag("div");
-        html.text(node.getChars());
-        context.renderChildren(node);
-        html.closeTag("div");
-    }
-
     public static class Factory implements NodeRendererFactory {
         @Override
         public NodeRenderer create(final DataHolder options) {
-            return new ConcordionNodeRenderer(options);
+            return new ConcordionCommandNodeRenderer(options);
         }
     }
 }
