+++ src/main/java/org/concordion/internal/parser/flexmark/ConcordionRunNode.java
@@ -3,20 +3,14 @@ package org.concordion.internal.parser.flexmark;
 import com.vladsch.flexmark.ast.LinkNode;
 import com.vladsch.flexmark.util.sequence.BasedSequence;
 
-public class ConcordionLinkNode extends LinkNode {
+public class ConcordionRunNode extends LinkNode {
 
     private String runner;
     private String command;
 
-    // TODO check what this does
     @Override
     public BasedSequence[] getSegments() {
         return EMPTY_SEGMENTS;
-    }
-
-    @Override
-    public void getAstExtra(StringBuilder out) {
-        astExtraChars(out);
     }
 
     public String getRunner() {
