+++ src/main/java/org/concordion/internal/parser/flexmark/FlexmarkConcordionExtension.java
@@ -1,0 +1,40 @@
+package org.concordion.internal.parser.flexmark;
+
+import com.vladsch.flexmark.Extension;
+import com.vladsch.flexmark.html.HtmlRenderer;
+import com.vladsch.flexmark.parser.Parser;
+import com.vladsch.flexmark.util.options.MutableDataHolder;
+
+public class FlexmarkConcordionExtension implements Parser.ParserExtension, HtmlRenderer.HtmlRendererExtension {
+    private FlexmarkConcordionExtension() {
+    }
+
+    public static Extension create() {
+        return new FlexmarkConcordionExtension();
+    }
+
+    @Override
+    public void extend(Parser.Builder parserBuilder) {
+        parserBuilder.postProcessorFactory(new ConcordionExampleNodePostProcessor.Factory(parserBuilder));
+        parserBuilder.postProcessorFactory(new ConcordionRunNodePostProcessor.Factory(parserBuilder));
+        parserBuilder.postProcessorFactory(new ConcordionCommandNodePostProcessor.Factory(parserBuilder));
+    }
+
+    @Override
+    public void rendererOptions(final MutableDataHolder options) {
+    }
+
+    @Override
+    public void parserOptions(final MutableDataHolder options) {
+
+    }
+
+    @Override
+    public void extend(HtmlRenderer.Builder rendererBuilder, String rendererType) {
+        if (rendererType.equals("HTML")) {
+            rendererBuilder.nodeRendererFactory(new ConcordionCommandNodeRenderer.Factory());
+            rendererBuilder.nodeRendererFactory(new ConcordionExampleNodeRenderer.Factory());
+            rendererBuilder.nodeRendererFactory(new ConcordionRunNodeRenderer.Factory());
+        }
+    }
+}
