+++ src/test/java/spec/concordion/specificationType/markdown/MarkdownGrammarFixture.java
@@ -3,18 +3,17 @@ package spec.concordion.specificationType.markdown;
 import java.util.HashMap;
 
 import org.concordion.integration.junit4.ConcordionRunner;
-import org.concordion.internal.parser.flexmark.FlexmarkParser;
-import org.concordion.internal.parser.markdown.MarkdownParser;
+import org.concordion.internal.parser.flexmark.MarkdownToHtml;
 import org.junit.runner.RunWith;
 
 @RunWith(ConcordionRunner.class)
 public class MarkdownGrammarFixture {
-    FlexmarkParser markdownParser;
+    MarkdownToHtml markdownParser;
 
     public MarkdownGrammarFixture() {
         HashMap<String, String> namespaces = new HashMap<String, String>();
         namespaces.put("ext", "urn:concordion-extensions:2010");
-        markdownParser = new FlexmarkParser(0, namespaces, "concordion");
+        markdownParser = new MarkdownToHtml(0, namespaces, "concordion");
     }
 
     public String translate(String markdown) {
